+++ batch/job_publisher.js
@@ -1,12 +1,32 @@
 'use strict';
 
-function JobPublisher(redis) {
+var debug = require('./util/debug')('pubsub:publisher');
+var error = require('./util/debug')('pubsub:publisher:error');
+
+var DB = 0;
+
+function JobPublisher(pool) {
+    this.pool = pool;
     this.channel = 'batch:hosts';
-    this.client = redis.createClient(global.settings.redis_port, global.settings.redis_host);
 }
 
 JobPublisher.prototype.publish = function (host) {
-    this.client.publish(this.channel, host);
+    var self = this;
+
+    this.pool.acquire(DB, function (err, client) {
+        if (err) {
+            return error('Error adquiring redis client: ' + err.message);
+        }
+
+        client.publish(self.channel, host, function (err) {
+            if (err) {
+                return error('Error publishing to ' + self.channel + ':' + host + ', ' + err.message);
+            }
+
+            debug('publish to ' + self.channel + ':' + host);
+            self.pool.release(DB, client);
+        });
+    });
 };
 
 module.exports = JobPublisher;