+++ batch/job_subscriber.js
@@ -1,24 +1,32 @@
 'use strict';
 
-var debug = require('./util/debug')('job-subscriber');
+var debug = require('./util/debug')('pubsub:subscriber');
+var error = require('./util/debug')('pubsub:subscriber:error');
+
+var DB = 0;
 var SUBSCRIBE_INTERVAL_IN_MILLISECONDS = 10 * 60 * 1000; // 10 minutes
 
 function _subscribe(client, channel, queueSeeker, onMessage) {
+
     queueSeeker.seek(onMessage, function (err) {
         if (err) {
-            debug(err);
+            error(err);
         }
 
         client.removeAllListeners('message');
         client.unsubscribe(channel);
         client.subscribe(channel);
-        client.on('message', onMessage);
+
+        client.on('message', function (channel, host) {
+            debug('message received from: ' + channel + ':' + host);
+            onMessage(channel, host);
+        });
     });
 }
 
-function JobSubscriber(redis, queueSeeker) {
+function JobSubscriber(pool, queueSeeker) {
     this.channel = 'batch:hosts';
-    this.client = redis.createClient(global.settings.redis_port, global.settings.redis_host);
+    this.pool = pool;
     this.queueSeeker = queueSeeker;
 }
 
@@ -27,19 +35,30 @@ module.exports = JobSubscriber;
 JobSubscriber.prototype.subscribe = function (onMessage) {
     var self = this;
 
-    _subscribe(this.client, this.channel, this.queueSeeker, onMessage);
+    this.pool.acquire(DB, function (err, client) {
+        if (err) {
+            return error('Error adquiring redis client: ' + err.message);
+        }
 
-    this.seekerInterval = setInterval(
-        _subscribe,
-        SUBSCRIBE_INTERVAL_IN_MILLISECONDS,
-        this.client,
-        this.channel,
-        self.queueSeeker,
-        onMessage
-    );
+        self.client = client;
+
+        _subscribe(self.client, self.channel, self.queueSeeker, onMessage);
+
+        self.seekerInterval = setInterval(
+            _subscribe,
+            SUBSCRIBE_INTERVAL_IN_MILLISECONDS,
+            self.client,
+            self.channel,
+            self.queueSeeker,
+            onMessage
+        );
+    });
+
 };
 
 JobSubscriber.prototype.unsubscribe = function () {
     clearInterval(this.seekerInterval);
-    this.client.unsubscribe(this.channel);
+    if (this.client && this.client.connected) {
+        this.client.unsubscribe(this.channel);
+    }
 };